+++ dbs-21/src/components/Itinerary.jsx
@@ -1,22 +1,17 @@
 import React from 'react'
 import Destination from './Destination'
 
-const Itinerary = ({itinerary, onDelete}) => {
+const Itinerary = ({itinerary, onDelete, onEdit}) => {
   return (
     <div className= 'itinerary'>
     <h1>Title: {itinerary.title} </h1>
     <h2>Budget: ${itinerary.budget}</h2>
      <h3> Destinations </h3>
-     <button> Add Destination</button>
     {itinerary.destinations.map((destination, index) => (
-        <Destination key={index} destination={destination} onDelete={onDelete} />
+        <Destination key={index} destination={destination} onDelete={onDelete} onEdit={onEdit} />
        ))} 
-       <button >
-        Edit
-      </button>
-      <button>
-        Delete
-      </button>
+       <button className="edit-button"> Edit Itinerary </button>
+      <button className= "delete-button"> Delete Itinerary </button>
     </div>
 
   )
